+++ src/Simulateur/Blockchain.java
@@ -9,9 +9,9 @@ public interface Blockchain extends Remote, Serializable{
     //Print a blockchain from a neighbour
     public StringBuilder printBlockchainImpl(String myPort) throws RemoteException ;
     //Ask a blockchain of a neighbour, from 0 to profondeur
-    public Blockchain askBlockchain(BigDecimal profondeur) throws RemoteException ;
+    public Blockchain askBlockchain(int profondeur) throws RemoteException ;
 
-    public BigDecimal getHeight() throws RemoteException ;
+    public int getHeight() throws RemoteException ;
 
     public LinkedList<Block> sendBlockList() throws RemoteException;
